# docs/openapi.yaml
openapi: 3.0.3
info:
  title: sandboxd /v1 API
  version: "0.3.0"
  description: >
    The public, tenant-scoped API for sandboxd. This is the only surface a
    console or external integration should use — the internal /sandbox and
    /wake endpoints are not part of this contract. Scoped by the API token
    (Bearer); when auth is disabled (the single-user default) the whole
    surface is reachable without a token.
servers:
  - url: http://127.0.0.1:9090
    description: Default loopback bind (docker-compose maps SANDBOXD_API_BIND here)
security:
  - bearerAuth: []
  - {} # auth disabled by default (single-user)
...
paths:
...
  /v1/apps/{id}/config:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      operationId: listAppConfig
      tags: [apps]
      summary: List an app's config entries (metadata; secret values redacted)
      description: >
        Returns every config key for the app. Sensitive entries never
        include their value (only value_set=true); non-sensitive entries
        include the plaintext value.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  config:
                    type: array
                    items: { $ref: "#/components/schemas/ConfigItem" }
        "404": { $ref: "#/components/responses/Error" }
    post:
      operationId: createAppConfig
      tags: [apps]
      summary: Create a config entry (set sensitive=true to encrypt at rest)
      description: >
        Sensitive values are AES-256-GCM-encrypted at rest and are
        redacted — they are never returned by any GET. access_policy
        governs who may later read the value through the broker; the
        default control_plane_only is not supplied to runtime processes.
        The owning tenant service may explicitly reveal one stored value
        using the separate POST reveal endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateConfigRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigItem" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409":
          description: A config entry with this key already exists; PATCH to update.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/apps/{id}/config/{key}:
    parameters:
      - $ref: "#/components/parameters/Id"
      - in: path
        name: key
        required: true
        schema: { type: string }
    patch:
      operationId: updateAppConfig
      tags: [apps]
      summary: Update a config entry (value, sensitivity, or access policy)
      description: >
        Omit value to keep the stored value untouched (a policy-only
        change preserves an existing secret). Toggling sensitive
        re-encodes the stored value.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchConfigRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigItem" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteAppConfig
      tags: [apps]
      summary: Delete a config entry
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    cubeRelayKey:
      type: apiKey
      in: header
      name: x-api-key
    cubeBridgeToken:
      type: apiKey
      in: header
      name: x-baarcha-bridge
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            retryable: { type: boolean }
...
    ConfigItem:
      type: object
      properties:
        key: { type: string }
        sensitive: { type: boolean }
        access_policy:
          type: string
          enum: [control_plane_only, agent_access, runtime_access, both]
        value_set: { type: boolean, description: "True when a value is stored." }
        value:
          type: string
          description: "Present only for non-sensitive entries; sensitive values are never returned."
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateConfigRequest:
      type: object
      required: [key]
      properties:
        key: { type: string }
        value: { type: string }
        sensitive: { type: boolean, default: false }
        access_policy:
          type: string
          enum: [control_plane_only, agent_access, runtime_access, both]
          default: control_plane_only
    PatchConfigRequest:
      type: object
      properties:
        value: { type: string, description: "Omit to keep the stored value unchanged." }
        sensitive: { type: boolean }
        access_policy:
          type: string
          enum: [control_plane_only, agent_access, runtime_access, both]
...